+++ Bonobo.Git.Server/HISTORY.md
@@ -1,6 +1,7 @@
 ﻿## 2019-07-24
 
 - Update LogFilger to log CurrentExecutionFilePath
+- Update Markdown render to `Markdig`
 
 ## 2019-04-04
 - Add NLog support